+++ src/main/groovy/com/cedarsoftware/ncube/NCubeMutableClient.groovy
@@ -149,4 +149,8 @@ interface NCubeMutableClient extends NCubeClient
     void createRefAxis(ApplicationID appId, String cubeName, String axisName, ApplicationID refAppId, String refCubeName, String refAxisName)
 
     NCubeInfoDto promoteRevision(long cubeId)
+
+    List<Delta> fetchJsonRevDiffs(long newCubeId, long oldCubeId)
+
+    List<Delta> fetchJsonBranchDiffs(NCubeInfoDto newInfoDto, NCubeInfoDto oldInfoDto)
 }